requirements.txt
ShopifyAPI
python-dotenv
openpyxl